+++ tanaguru-engine/tanaguru-api/src/main/java/com/tanaguru/service/impl/AuditParameterServiceImpl.java
@@ -3,6 +3,7 @@ package com.tanaguru.service.impl;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.tanaguru.config.PropertyConfig;
+import com.tanaguru.domain.constant.CustomError;
 import com.tanaguru.domain.constant.EAuditParameter;
 import com.tanaguru.domain.constant.EAuditType;
 import com.tanaguru.domain.constant.EParameterFamily;
@@ -12,7 +13,7 @@ import com.tanaguru.domain.entity.audit.parameter.AuditParameter;
 import com.tanaguru.domain.entity.audit.parameter.AuditParameterFamily;
 import com.tanaguru.domain.entity.audit.parameter.AuditParameterValue;
 import com.tanaguru.domain.entity.membership.project.Project;
-import com.tanaguru.domain.exception.InvalidEntityException;
+import com.tanaguru.domain.exception.CustomInvalidEntityException;
 import com.tanaguru.helper.AESEncrypt;
 import com.tanaguru.helper.UrlHelper;
 import com.tanaguru.repository.*;
@@ -130,7 +131,7 @@ public class AuditParameterServiceImpl implements AuditParameterService {
             auditParameterDefaultValueMap.put(
                     auditParameter.getCode(),
                     auditParameterValueRepository.findFirstByIsDefaultAndAuditParameter(true, auditParameter)
-                            .orElseThrow(() -> new InvalidEntityException("Audit parameter " + auditParameter.getCode() + "has not default value"))
+                            .orElseThrow(() -> new CustomInvalidEntityException(CustomError.AUDIT_PARAMETER_NO_DEFAULT_VALUE , auditParameter.getCode().toString() ))
             );
         }
     }
@@ -196,7 +197,7 @@ public class AuditParameterServiceImpl implements AuditParameterService {
 
                 definiteParameters.put(parameter, existingParameterValueOpt);
             } else {
-                throw new InvalidEntityException("Invalid value for parameter " + parameter.getCode() + " : " + value);
+                throw new CustomInvalidEntityException(CustomError.INVALID_VALUE_PARAMETER, parameter.getCode().toString(), value );
             }
         }