+++ tanaguru-engine/tanaguru-api/src/test/java/factory/impl/UserFactoryTest.java
@@ -3,6 +3,7 @@ package factory.impl;
 import com.tanaguru.domain.constant.EAppRole;
 import com.tanaguru.domain.entity.membership.user.AppRole;
 import com.tanaguru.domain.entity.membership.user.User;
+import com.tanaguru.domain.exception.CustomEntityNotFoundException;
 import com.tanaguru.factory.impl.UserFactoryImpl;
 import com.tanaguru.repository.ContractRepository;
 import com.tanaguru.repository.UserRepository;
@@ -15,7 +16,6 @@ import org.mockito.Mock;
 import org.mockito.Mockito;
 import org.mockito.junit.MockitoJUnitRunner;
 
-import javax.persistence.EntityNotFoundException;
 import java.util.Optional;
 
 @RunWith(MockitoJUnitRunner.class)
@@ -32,7 +32,7 @@ public class UserFactoryTest {
     @InjectMocks
     private UserFactoryImpl userFactory;
 
-    @Test(expected = EntityNotFoundException.class)
+    @Test(expected = CustomEntityNotFoundException.class)
     public void createUserTest_RoleNotFound() {
         Mockito.when(appRoleService.getAppRole(EAppRole.ADMIN)).thenReturn(Optional.empty());
         userFactory.createUser("test", "test@test.com", "test", EAppRole.ADMIN, false);