+++ tanaguru-engine/tanaguru-audit-runner/src/main/java/com/tanaguru/runner/factory/AuditRunnerFactoryImpl.java
@@ -3,11 +3,13 @@ package com.tanaguru.runner.factory;
 import com.tanaguru.config.PropertyConfig;
 import com.tanaguru.crawler.TanaguruCrawlerController;
 import com.tanaguru.crawler.factory.TanaguruCrawlerControllerFactory;
+import com.tanaguru.domain.constant.CustomError;
 import com.tanaguru.domain.constant.BrowserName;
 import com.tanaguru.domain.constant.EAuditLogLevel;
 import com.tanaguru.domain.constant.EAuditParameter;
 import com.tanaguru.domain.entity.audit.*;
 import com.tanaguru.domain.entity.audit.parameter.AuditParameterValue;
+import com.tanaguru.domain.exception.CustomEntityNotFoundException;
 import com.tanaguru.driver.factory.TanaguruDriverFactory;
 import com.tanaguru.helper.AESEncrypt;
 import com.tanaguru.repository.AuditReferenceRepository;
@@ -22,7 +24,6 @@ import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
-import javax.persistence.EntityNotFoundException;
 import javax.transaction.Transactional;
 import java.util.*;
 import java.util.stream.Collectors;
@@ -146,7 +147,7 @@ public class AuditRunnerFactoryImpl implements AuditRunnerFactory {
             case SCENARIO:
                 long scenarioId = Long.parseLong(parameterStringMap.get(EAuditParameter.SCENARIO_ID).getValue());
                 Scenario scenario = scenarioRepository.findById(scenarioId)
-                        .orElseThrow(() -> new EntityNotFoundException("Cannot find scenario with id " + scenarioId));
+                        .orElseThrow(() -> new CustomEntityNotFoundException(CustomError.SCENARIO_NOT_FOUND, scenarioId ));
 
                 result = createSeleneseRunner(
                         tanaguruTests,
@@ -163,7 +164,7 @@ public class AuditRunnerFactoryImpl implements AuditRunnerFactory {
             case UPLOAD:
                 long resourceId = Long.parseLong(parameterStringMap.get(EAuditParameter.DOM_ID).getValue());
                 Resource resource = resourceRepository.findById(resourceId)
-                        .orElseThrow(() -> new EntityNotFoundException("Cannot find resource with id " + resourceId));
+                        .orElseThrow(() -> new CustomEntityNotFoundException(CustomError.RESOURCE_NOT_FOUND, resourceId ));
                 result = createFileRunner(
                         tanaguruTests,
                         audit,