+++ tanaguru-rest/src/main/java/com/tanaguru/controller/ConfigurationController.java
@@ -1,7 +1,10 @@
 package com.tanaguru.controller;
 
 import io.swagger.annotations.ApiOperation;
+
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
+import org.springframework.boot.info.BuildProperties;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseBody;
@@ -20,6 +23,9 @@ public class ConfigurationController {
     @Value("${auditrunner.active}")
     private String[] browsers;
 
+    @Autowired
+    BuildProperties buildProperties;
+    
     @ApiOperation(
             value = "Get session duration value",
             response = Long.class
@@ -38,6 +44,16 @@ public class ConfigurationController {
     public @ResponseBody
     String[] getBrowsersEnabled(){
         return this.browsers;
+    }
+    
+    @ApiOperation(
+            value = "Get engine version",
+            response = String.class
+    )
+    @GetMapping("/engine-version")
+    public @ResponseBody
+    String getEngineVersion(){
+        return this.buildProperties.getVersion();
     }
     
 }