+++ tanaguru-rest/src/main/java/com/tanaguru/controller/PageController.java
@@ -1,5 +1,8 @@
 package com.tanaguru.controller;
 
+import com.tanaguru.domain.constant.CustomError;
+import com.tanaguru.domain.exception.CustomEntityNotFoundException;
+import com.tanaguru.domain.exception.CustomForbiddenException;
 import com.tanaguru.domain.entity.audit.Page;
 import com.tanaguru.domain.exception.ForbiddenException;
 import com.tanaguru.helper.JsonHttpHeaderBuilder;
@@ -22,7 +25,6 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
 
-import javax.persistence.EntityNotFoundException;
 import java.util.Collection;
 
 /**
@@ -51,11 +53,14 @@ public class PageController {
     @ApiOperation(
             value = "Get Page for a given id",
             notes = "User must have SHOW_AUDIT authority on page's project or a valid sharecode"
+                    + "\nIf page not found, exception raise : PAGE_NOT_FOUND with page id"
+                    + "\nIf user cannot access page content, exception raise : CANNOT_ACCESS_PAGE_CONTENT_FOR_PAGE with page id"
     )
     @ApiResponses(value = {
             @ApiResponse(code = 400, message = "Invalid parameters"),
             @ApiResponse(code = 403, message = "Forbidden for current session or invalid sharecode"),
-            @ApiResponse(code = 404, message = "Audit not found")
+            @ApiResponse(code = 404, message = "Page not found : PAGE_NOT_FOUND error"
+                    + "\nCannot access page content for page : CANNOT_ACCESS_PAGE_CONTENT_FOR_PAGE error")
     })
     @GetMapping("/{id}/{shareCode}")
     public @ResponseBody
@@ -63,12 +68,12 @@ public class PageController {
             @PathVariable long id,
             @PathVariable(required = false) @ApiParam(required = false) String shareCode) {
         Page page = pageRepository.findById(id)
-                .orElseThrow(EntityNotFoundException::new);
+                .orElseThrow(() -> new CustomEntityNotFoundException(CustomError.PAGE_NOT_FOUND, id ));
 
         if(tanaguruUserDetailsService.currentUserCanShowAudit(page.getAudit().getId(), shareCode)){
             return page;
         }else{
-            throw new ForbiddenException("Cannot access page " + id);
+            throw new CustomForbiddenException(CustomError.CANNOT_ACCESS_PAGE_CONTENT_FOR_PAGE, id );
         }
     }
 
@@ -82,11 +87,12 @@ public class PageController {
     @ApiOperation(
             value = "Get all Page for a given Audit id",
             notes = "User must have SHOW_AUDIT authority on audit's project or a valid sharecode"
+                    + "\nIf user cannot access pages for the audit, exception raise : CANNOT_ACCESS_PAGES_FOR_AUDIT with audit id"
     )
     @ApiResponses(value = {
             @ApiResponse(code = 400, message = "Invalid parameters"),
             @ApiResponse(code = 403, message = "Forbidden for current session or invalid sharecode"),
-            @ApiResponse(code = 404, message = "Audit not found")
+            @ApiResponse(code = 404, message = "Cannot access pages for audit : CANNOT_ACCESS_PAGES_FOR_AUDIT error")
     })
     @GetMapping("/by-audit/{id}/{shareCode}")
     public @ResponseBody
@@ -96,7 +102,7 @@ public class PageController {
         if(tanaguruUserDetailsService.currentUserCanShowAudit(id, shareCode)){
             return pageRepository.findAllByAudit_Id(id);
         }else{
-            throw new ForbiddenException("Cannot access pages for audit " + id);
+            throw new CustomForbiddenException(CustomError.CANNOT_ACCESS_PAGES_FOR_AUDIT, id );
         }
     }
 
@@ -110,11 +116,12 @@ public class PageController {
     @ApiOperation(
             value = "Get paginated Page for a given Audit id",
             notes = "User must have SHOW_AUDIT authority on audit's project or a valid sharecode"
+                    + "\nIf user cannot access pages for the audit, exception raise : CANNOT_ACCESS_PAGES_FOR_AUDIT"
     )
     @ApiResponses(value = {
             @ApiResponse(code = 400, message = "Invalid parameters"),
             @ApiResponse(code = 403, message = "Forbidden for current session or invalid sharecode"),
-            @ApiResponse(code = 404, message = "Audit not found")
+            @ApiResponse(code = 404, message = "Cannot access pages for audit : CANNOT_ACCESS_PAGES_FOR_AUDIT error")
     })
     @GetMapping("/by-audit-paginated/{id}/{shareCode}")
     public @ResponseBody
@@ -126,7 +133,7 @@ public class PageController {
         if(tanaguruUserDetailsService.currentUserCanShowAudit(id, shareCode)){
             return pageRepository.findAllByAudit_Id(id, PageRequest.of(page, size));
         }else{
-            throw new ForbiddenException("Cannot access pages for audit " + id);
+            throw new CustomForbiddenException(CustomError.CANNOT_ACCESS_PAGES_FOR_AUDIT, id );
         }
     }