+++ tanaguru-rest/src/main/java/com/tanaguru/controller/TanaguruTestController.java
@@ -1,9 +1,11 @@
 package com.tanaguru.controller;
 
+import com.tanaguru.domain.constant.CustomError;
 import com.tanaguru.domain.dto.TanaguruTestDTO;
 import com.tanaguru.domain.entity.audit.Audit;
 import com.tanaguru.domain.entity.audit.TanaguruTest;
 import com.tanaguru.domain.entity.audit.TestHierarchy;
+import com.tanaguru.domain.exception.CustomEntityNotFoundException;
 import com.tanaguru.repository.TanaguruTestRepository;
 import com.tanaguru.repository.TestHierarchyRepository;
 import io.swagger.annotations.ApiOperation;
@@ -12,7 +14,6 @@ import io.swagger.annotations.ApiResponses;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
 
-import javax.persistence.EntityNotFoundException;
 import java.util.Collection;
 import java.util.stream.Collectors;
 
@@ -36,10 +37,11 @@ public class TanaguruTestController {
      * @return The @see TanaguruTest
      */
     @ApiOperation(
-            value = "Get TanaguruTest for a given id")
+            value = "Get TanaguruTest for a given id",
+            notes = "If tanaguru test not found, exception raise : TANAGURU_TEST_NOT_FOUND with tanaguru test id")
     @ApiResponses(value = {
             @ApiResponse(code = 400, message = "Invalid parameters"),
-            @ApiResponse(code = 404, message = "TanaguruTest not found")
+            @ApiResponse(code = 404, message = "TanaguruTest not found : TANAGURU_TEST_NOT_FOUND error")
     })
     @GetMapping("/{id}")
     public @ResponseBody
@@ -47,7 +49,7 @@ public class TanaguruTestController {
             @PathVariable long id) {
         return new TanaguruTestDTO(
                 tanaguruTestRepository.findById(id)
-                    .orElseThrow(EntityNotFoundException::new));
+                    .orElseThrow(() -> new CustomEntityNotFoundException(CustomError.TANAGURU_TEST_NOT_FOUND, id )));
     }
 
     /**
@@ -56,17 +58,18 @@ public class TanaguruTestController {
      * @return The collection of @see TanaguruTest
      */
     @ApiOperation(
-            value = "Get all TanaguruTest for a given referernce id")
+            value = "Get all TanaguruTest for a given reference id",
+            notes = "If test hierarchy not found, exception raise : TEST_HIERARCHY_NOT_FOUND with id")
     @ApiResponses(value = {
             @ApiResponse(code = 400, message = "Invalid parameters"),
-            @ApiResponse(code = 404, message = "Reference not found")
+            @ApiResponse(code = 404, message = "Test hierarchy not found : TEST_HIERARCHY_NOT_FOUND error")
     })
     @GetMapping("/by-reference/{id}")
     public @ResponseBody
     Collection<TanaguruTestDTO> getByReferenceId(
             @PathVariable long id) {
         TestHierarchy reference = testHierarchyRepository.findByIdAndIsDeletedIsFalseAndParentIsNull(id)
-                .orElseThrow(EntityNotFoundException::new);
+                .orElseThrow(() -> new CustomEntityNotFoundException(CustomError.TEST_HIERARCHY_NOT_FOUND, id ));
 
         return tanaguruTestRepository.findAllByTestHierarchies_ReferenceAndIsDeletedIsFalse(reference)
                 .stream()
@@ -80,17 +83,18 @@ public class TanaguruTestController {
      * @return The collection of @see TanaguruTest
      */
     @ApiOperation(
-            value = "Get all TanaguruTest for a given TestHierarchy id")
+            value = "Get all TanaguruTest for a given TestHierarchy id",
+            notes = "If test hierarchy not found, exception raise : TEST_HIERARCHY_NOT_FOUND with id")
     @ApiResponses(value = {
             @ApiResponse(code = 400, message = "Invalid parameters"),
-            @ApiResponse(code = 404, message = "TestHierarchy not found")
+            @ApiResponse(code = 404, message = "Test hierarchy not found : TEST_HIERARCHY_NOT_FOUND error")
     })
     @GetMapping("/by-test-hierarchy/{id}")
     public @ResponseBody
     Collection<TanaguruTestDTO> getByTestHierarchyId(
             @PathVariable long id) {
         TestHierarchy testHierarchy = testHierarchyRepository.findById(id)
-                .orElseThrow(EntityNotFoundException::new);
+                .orElseThrow(() -> new CustomEntityNotFoundException(CustomError.TEST_HIERARCHY_NOT_FOUND, id ));
 
         return tanaguruTestRepository.findAllByTestHierarchiesContainsAndIsDeletedIsFalse(testHierarchy)
                 .stream()
@@ -107,7 +111,7 @@ public class TanaguruTestController {
             notes = "User must have CREATE_TEST authority")
     @ApiResponses(value = {
             @ApiResponse(code = 400, message = "Invalid parameters"),
-            @ApiResponse(code = 401, message = "Unauthorized"),
+            @ApiResponse(code = 401, message = "Unauthorized : ACCESS_DENIED message"),
             @ApiResponse(code = 403, message = "Forbidden for current session")
     })
     @PreAuthorize(