+++ tanaguru-rest/src/main/java/com/tanaguru/controller/TestHierarchyController.java
@@ -1,10 +1,13 @@
 package com.tanaguru.controller;
 
+import com.tanaguru.domain.constant.CustomError;
+import com.tanaguru.domain.exception.CustomEntityNotFoundException;
+import com.tanaguru.domain.exception.CustomForbiddenException;
+import com.tanaguru.domain.exception.CustomInvalidArgumentException;
 import com.tanaguru.domain.dto.TestHierarchyDTO;
 import com.tanaguru.domain.entity.audit.Audit;
 import com.tanaguru.domain.entity.audit.TanaguruTest;
 import com.tanaguru.domain.entity.audit.TestHierarchy;
-import com.tanaguru.domain.exception.ForbiddenException;
 import com.tanaguru.repository.AuditRepository;
 import com.tanaguru.repository.TanaguruTestRepository;
 import com.tanaguru.repository.TestHierarchyRepository;
@@ -14,12 +17,10 @@ import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiParam;
 import io.swagger.annotations.ApiResponse;
 import io.swagger.annotations.ApiResponses;
-import org.openqa.selenium.InvalidArgumentException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
 
-import javax.persistence.EntityNotFoundException;
 import javax.validation.Valid;
 import java.util.Collection;
 import java.util.Comparator;
@@ -53,17 +54,18 @@ public class TestHierarchyController {
      * @return The @see TestHierarchy
      */
     @ApiOperation(
-            value = "Get TestHierarchy for a given id")
+            value = "Get TestHierarchy for a given id",
+            notes = "If test hierarchy not found, exception raise : TEST_HIERARCHY_NOT_FOUND with test hierarchy id")
     @ApiResponses(value = {
             @ApiResponse(code = 400, message = "Invalid parameters"),
-            @ApiResponse(code = 404, message = "TestHierarchy not found")
+            @ApiResponse(code = 404, message = "TestHierarchy not found : TEST_HIERARCHY_NOT_FOUND error")
     })
     @GetMapping("/{id}")
     public @ResponseBody
     TestHierarchyDTO getById(
             @PathVariable long id) {
         return new TestHierarchyDTO(testHierarchyRepository.findById(id)
-                .orElseThrow(EntityNotFoundException::new));
+                .orElseThrow(() -> new CustomEntityNotFoundException(CustomError.TEST_HIERARCHY_NOT_FOUND, id )));
     }
 
     /**
@@ -106,17 +108,18 @@ public class TestHierarchyController {
      * @return The collection of @see TestHierarchy
      */
     @ApiOperation(
-            value = "Get all TestHierarchy children of a given TestHierarchy id")
+            value = "Get all TestHierarchy children of a given TestHierarchy id",
+            notes = "If test hierarchy not found, exception raise : TEST_HIERARCHY_NOT_FOUND with test hierarchy id")
     @ApiResponses(value = {
             @ApiResponse(code = 400, message = "Invalid parameters"),
-            @ApiResponse(code = 404, message = "TestHierarchy not found")
+            @ApiResponse(code = 404, message = "TestHierarchy not found : TEST_HIERARCHY_NOT_FOUND error")
     })
     @GetMapping("by-parent/{id}")
     public @ResponseBody
     Collection<TestHierarchyDTO> getByParentId(
             @PathVariable long id) {
         return testHierarchyRepository.findById(id)
-                .orElseThrow(EntityNotFoundException::new)
+                .orElseThrow(() -> new CustomEntityNotFoundException(CustomError.TEST_HIERARCHY_NOT_FOUND, id ))
                 .getChildren().stream()
                 .sorted(Comparator.comparingInt(TestHierarchy::getRank))
                 .map(TestHierarchyDTO::new)
@@ -124,10 +127,13 @@ public class TestHierarchyController {
     }
 
     @ApiOperation(
-            value = "Get all TestHierarchy for a given TanaguruTest id and TestHierarchy id")
+            value = "Get all TestHierarchy for a given TanaguruTest id and TestHierarchy id",
+            notes = "If test hierarchy not found, exception raise : TEST_HIERARCHY_NOT_FOUND with test hierarchy id"
+                    + "\nIf tanaguru test not found, exception raise : TANAGURU_TEST_NOT_FOUND with tanaguru test id")
     @ApiResponses(value = {
             @ApiResponse(code = 400, message = "Invalid parameters"),
-            @ApiResponse(code = 404, message = "TanaguruTest or TestHierarchy not found")
+            @ApiResponse(code = 404, message = "TestHierarchy not found : TEST_HIERARCHY_NOT_FOUND error"
+                    + "\nTanaguru test not found : TANAGURU_TEST_NOT_FOUND error")
     })
     @GetMapping("by-test-and-reference/{testId}/{referenceId}")
     public @ResponseBody
@@ -135,10 +141,10 @@ public class TestHierarchyController {
             @PathVariable long testId,
             @PathVariable long referenceId) {
         TestHierarchy reference = testHierarchyRepository.findById(referenceId)
-                .orElseThrow(EntityNotFoundException::new);
+                .orElseThrow(() -> new CustomEntityNotFoundException(CustomError.TEST_HIERARCHY_NOT_FOUND, referenceId ));
 
         TanaguruTest test = tanaguruTestRepository.findById(testId)
-                .orElseThrow(() -> new EntityNotFoundException("Cannot find Tanaguru test with id " + testId));
+                .orElseThrow(() -> new CustomEntityNotFoundException(CustomError.TANAGURU_TEST_NOT_FOUND, testId ));
 
         return testHierarchyRepository.findByReferenceAndTanaguruTestsContains(reference, test)
                 .stream().map(TestHierarchyDTO::new).collect(Collectors.toList());
@@ -151,12 +157,15 @@ public class TestHierarchyController {
      */
     @ApiOperation(
             value = "Get all references for a given audit id",
-            notes = "User must have SHOW_AUDIT authority on project or a valid sharecode")
+            notes = "User must have SHOW_AUDIT authority on project or a valid sharecode"
+                    + "\nIf audit not found, exception raise : AUDIT_NOT_FOUND with audit id"
+                    + "\nIf cannot show audit, exception raise : CANNOT_SHOW_AUDIT with audit id")
     @ApiResponses(value = {
             @ApiResponse(code = 400, message = "Invalid parameters"),
-            @ApiResponse(code = 401, message = "Unauthorized"),
+            @ApiResponse(code = 401, message = "Unauthorized : ACCESS_DENIED message"),
             @ApiResponse(code = 403, message = "Forbidden for current session or invalid sharecode"),
-            @ApiResponse(code = 404, message = "Audit not found")
+            @ApiResponse(code = 404, message = "Audit not found : AUDIT_NOT_FOUND error"
+                    + "\nCannot show audit : CANNOT_SHOW_AUDIT error")
     })
     @GetMapping("/by-audit/{id}/{sharecode}")
     public @ResponseBody
@@ -164,10 +173,10 @@ public class TestHierarchyController {
             @PathVariable long id,
             @ApiParam(required = false) @PathVariable(required = false) String sharecode) {
         Audit audit = auditRepository.findById(id)
-                .orElseThrow(() -> new EntityNotFoundException("Cannot find audit with id " + id));
+                .orElseThrow(() -> new CustomEntityNotFoundException(CustomError.AUDIT_NOT_FOUND, id ));
 
         if(!tanaguruUserDetailsService.currentUserCanShowAudit(audit, sharecode)){
-            throw new ForbiddenException("Cannot show audit " + audit.getId());
+            throw new CustomForbiddenException(CustomError.CANNOT_SHOW_AUDIT, audit.getId() );
         }
 
         return audit.getAuditReferences().stream().map(
@@ -181,11 +190,15 @@ public class TestHierarchyController {
      */
     @ApiOperation(
             value = "Create a TestHierarchy",
-            notes = "User must have CREATE_REFERENCE authority")
+            notes = "User must have CREATE_REFERENCE authority"
+                    + "\nIf try to create multiples references, exception raise : CANNOT_CREATE_MULTIPLE_REFERENCES with test hierarchy code"
+                    + "\nIf test hierarchy not found, exception raise : TEST_HIERARCHY_NOT_FOUND with test hierarchy id")
     @ApiResponses(value = {
             @ApiResponse(code = 400, message = "Invalid parameters"),
-            @ApiResponse(code = 401, message = "Unauthorized"),
-            @ApiResponse(code = 403, message = "Forbidden for current session")
+            @ApiResponse(code = 401, message = "Unauthorized : ACCESS_DENIED message"),
+            @ApiResponse(code = 403, message = "Forbidden for current session"),
+            @ApiResponse(code = 404, message = "Test hierarchy not found : TEST_HIERARCHY_NOT_FOUND error"
+                    + "\nCannot create multiple references : CANNOT_CREATE_MULTIPLE_REFERENCES error")
     })
     @PreAuthorize(
             "hasAuthority(T(com.tanaguru.domain.constant.AppAuthorityName).CREATE_REFERENCE)")
@@ -193,7 +206,7 @@ public class TestHierarchyController {
     public @ResponseBody
     TestHierarchyDTO createTestHierarchy(@RequestBody @Valid TestHierarchyDTO testHierarchyDTO) {
         if(testHierarchyDTO.getParentId() != null && testHierarchyRepository.findByCodeAndParentIsNull(testHierarchyDTO.getCode()).isPresent()){
-            throw new InvalidArgumentException("Cannot create multiple references with same code " + testHierarchyDTO.getCode());
+            throw new CustomInvalidArgumentException(CustomError.CANNOT_CREATE_MULTIPLE_REFERENCES, testHierarchyDTO.getCode() );
         }
 
         TestHierarchy testHierarchy = new TestHierarchy();
@@ -206,14 +219,14 @@ public class TestHierarchyController {
                 testHierarchyDTO.getReferenceId() == null ?
                         null :
                         testHierarchyRepository.findById(testHierarchyDTO.getReferenceId())
-                                .orElseThrow(() -> new InvalidArgumentException("Cannot find test Hierarchy with id " + testHierarchyDTO.getReferenceId()))
+                                .orElseThrow(() -> new CustomInvalidArgumentException(CustomError.TEST_HIERARCHY_NOT_FOUND, testHierarchyDTO.getReferenceId() ))
         );
 
         testHierarchy.setParent(
                 testHierarchyDTO.getParentId() == null ?
                         null :
                         testHierarchyRepository.findById(testHierarchyDTO.getParentId())
-                                .orElseThrow(() -> new InvalidArgumentException("Cannot find test Hierarchy with id " + testHierarchyDTO.getParentId()))
+                                .orElseThrow(() -> new CustomInvalidArgumentException(CustomError.TEST_HIERARCHY_NOT_FOUND, testHierarchyDTO.getParentId() ))
                 );
 
         return new TestHierarchyDTO(testHierarchyRepository.save(testHierarchy));
@@ -224,12 +237,13 @@ public class TestHierarchyController {
      */
     @ApiOperation(
             value = "Delete a TestHierarchy and children",
-            notes = "User must have DELETE_REFERENCE authority")
+            notes = "User must have DELETE_REFERENCE authority"
+                    + "\nIf test hierarchy not found, exception raise : TEST_HIERARCHY_NOT_FOUND with test hierarchy id")
     @ApiResponses(value = {
             @ApiResponse(code = 400, message = "Invalid parameters"),
-            @ApiResponse(code = 401, message = "Unauthorized"),
+            @ApiResponse(code = 401, message = "Unauthorized : ACCESS_DENIED message"),
             @ApiResponse(code = 403, message = "Forbidden for current session"),
-            @ApiResponse(code = 404, message = "TestHierarchy not found")
+            @ApiResponse(code = 404, message = "TestHierarchy not found : TEST_HIERARCHY_NOT_FOUND error")
     })
     @PreAuthorize(
             "hasAuthority(T(com.tanaguru.domain.constant.AppAuthorityName).DELETE_REFERENCE)")
@@ -237,7 +251,7 @@ public class TestHierarchyController {
     public @ResponseBody
     void deleteReference(@PathVariable long id) {
         TestHierarchy testHierarchy = testHierarchyRepository.findByIdAndIsDeletedIsFalseAndParentIsNull(id)
-                .orElseThrow(() -> new EntityNotFoundException("Cannot find reference with id " + id));
+                .orElseThrow(() -> new CustomEntityNotFoundException(CustomError.TEST_HIERARCHY_NOT_FOUND, id ));
 
         testHierarchyService.deleteReference(testHierarchy);
     }
@@ -248,12 +262,15 @@ public class TestHierarchyController {
      */
     @ApiOperation(
             value = "Add a TanaguruTest to a given TestHierarchy",
-            notes = "User must have CREATE_REFERENCE authority")
+            notes = "User must have CREATE_REFERENCE authority"
+                    + "\nIf test hierarchy not found, exception raise : TEST_HIERARCHY_NOT_FOUND with test hierarchy id"
+                    + "\nIf tanaguru test not found, exception raise : TANAGURU_TEST_NOT_FOUND with tanaguru test id")
     @ApiResponses(value = {
             @ApiResponse(code = 400, message = "Invalid parameters"),
-            @ApiResponse(code = 401, message = "Unauthorized"),
+            @ApiResponse(code = 401, message = "Unauthorized : ACCESS_DENIED message"),
             @ApiResponse(code = 403, message = "Forbidden for current session"),
-            @ApiResponse(code = 404, message = "TestHierarchy or Tanaguru not found")
+            @ApiResponse(code = 404, message = "Test hierarchy not found : TEST_HIERARCHY_NOT_FOUND error"
+                    + "\nTanaguru test not found : TANAGURU_TEST_NOT_FOUND error")
     })
     @PreAuthorize(
             "hasAuthority(T(com.tanaguru.domain.constant.AppAuthorityName).CREATE_REFERENCE)")
@@ -261,10 +278,10 @@ public class TestHierarchyController {
     public @ResponseBody
     void addTestToTestHierarchy(@PathVariable long testHierarchyId, @PathVariable long testId) {
         TestHierarchy testHierarchy = testHierarchyRepository.findById(testHierarchyId)
-                .orElseThrow(() -> new EntityNotFoundException("Cannot find testHierarchy with id " + testHierarchyId));
+                .orElseThrow(() -> new CustomEntityNotFoundException(CustomError.TEST_HIERARCHY_NOT_FOUND, testHierarchyId ));
 
         TanaguruTest tanaguruTest = tanaguruTestRepository.findById(testId)
-                .orElseThrow(() -> new EntityNotFoundException("Cannot find test with id " + testId));
+                .orElseThrow(() -> new CustomEntityNotFoundException(CustomError.TANAGURU_TEST_NOT_FOUND, testId ));
 
         Collection<TanaguruTest> tanaguruTests = testHierarchy.getTanaguruTests();
         tanaguruTests.add(tanaguruTest);
@@ -277,12 +294,13 @@ public class TestHierarchyController {
      */
     @ApiOperation(
             value = "Add a TanaguruTest list to a given TestHierarchy",
-            notes = "User must have CREATE_REFERENCE authority")
+            notes = "User must have CREATE_REFERENCE authority"
+                    + "\nIf test hierarchy not found, exception raise : TEST_HIERARCHY_NOT_FOUND with test hierarchy id")
     @ApiResponses(value = {
             @ApiResponse(code = 400, message = "Invalid parameters"),
-            @ApiResponse(code = 401, message = "Unauthorized"),
+            @ApiResponse(code = 401, message = "Unauthorized : ACCESS_DENIED message"),
             @ApiResponse(code = 403, message = "Forbidden for current session"),
-            @ApiResponse(code = 404, message = "TestHierarchy or Tanaguru not found")
+            @ApiResponse(code = 404, message = "Test hierarchy not found : TEST_HIERARCHY_NOT_FOUND error")
     })
     @PreAuthorize(
             "hasAuthority(T(com.tanaguru.domain.constant.AppAuthorityName).CREATE_REFERENCE)")
@@ -290,7 +308,7 @@ public class TestHierarchyController {
     public @ResponseBody
     void addTestListToTestHierarchy(@PathVariable long testHierarchyId, @RequestBody Collection<Long> tanaguruTestIds) {
         TestHierarchy testHierarchy = testHierarchyRepository.findById(testHierarchyId)
-                .orElseThrow(() -> new EntityNotFoundException("Cannot find testHierarchy with id " + testHierarchyId));
+                .orElseThrow(() -> new CustomEntityNotFoundException(CustomError.TEST_HIERARCHY_NOT_FOUND, testHierarchyId ));
 
         Collection<TanaguruTest> tanaguruTests = tanaguruTestRepository.findAllById(tanaguruTestIds);