+++ CHANGELOG.md
@@ -3,4 +3,8 @@
 - Rest API
 - 4 types of audit
 - User, contract and project management
+
+## 1.0.2
+- Endpoint healthcheck , new key in common properties : management.endpoint.health.show-details, in order to show details or not
+- Rest : New key "admin.mail.whenblocked" (Allow send an email to admin when user is blocked)
 - JSON export of audits and pages